Initial Driver program and first attempt to get a ciTypeFlow
Driver application takes a class name and reflectively looks up the
`main(String[])` method.  It then passes this method to the native
code as the starting point for any analysis.

The goal is to get this program to the point where we can generate
a `ciTypeFlow` for the method and use that to determine with other
classes and methods are in use by the current method.

This currently crashes due to calls being done a java thread(?) that
should be happening on a CompilerThread.

Signed-off-by: Dan Heidinga <heidinga@redhat.com>