Create a special analysis thread and queue
ci must be used on a CompilerThread so create a new CompilerThread
that has a special entry point for our hacking.  It pulls tasks
from the _analysis_queue and then attempts to generate a ciTypeFlow
from the method.

There's some really hacky toNative and toVM transitions that seem
to be required here which makes me think the abstraction is wrong
but its enough to get the ciTypeFlow starting to be created.

Currently fails due to too many Nodes.  May be this can be worked
around by using Directives?

Signed-off-by: Dan Heidinga <heidinga@redhat.com>