Sketching out a {class,method,field} finder based on bcEscapeAnalyzer
Copying the approach in bcEscapeAnalyzer to walk the set of basic
blocks and process their bytecodes.  Will produce a set of further
methods to process, plus facts (RTA) about the classes detected in
the analysis.

Need to find a way to reflect this info back up into Java for the
Driver to track.

Also need to think about how to handle classes that aren't yet loaded
as that may result in simplified graphes due to assumptions during
the compilation process.  Likely depends on when / where we hook in?

Or maybe do a pre-pass over the bytecodes to discover the classes
and attempt to load them?

Signed-off-by: Dan Heidinga <heidinga@redhat.com>