HACK: allow compiler threads to call java code
Used to allow the StaticAnalyzer to load classes.  This should be
safe as - at least as far as I can tell - I'm doing the same things
to create the StaticAnalyzer's compile thread as JVMCI is doing.

My thread has a proper thread oop which seems to be the only necessary
thing (TDB)?

Signed-off-by: Dan Heidinga <heidinga@redhat.com>