StaticAnalyzer now processes a method at a time based on bcEscapeAnalyzer
The idea is to walk the bytecodes of each of the relevant (as discovered by
the compiler) blocks in the method.  For each class referenced in one of the
bytecodes: {
* ldc
* get/putstatic: field holder class
* invoke*: holder class of method being invoked
* anewarray: the klass of the array
* multianewarray: the klass of the array, assumes the processor will walk all dimensions to the root component
* instanceof/checkcast: the target class
}
or in the signature, we will force them to be loaded.

This is probably too aggressive in that it will load more classes than may strictly be
required, but I'm hoping we can remove them later.

My concern is that we may not see all relevant blocks due to unloaded
classes preventing paths from being accessible at the time we visit
the method for the first time.  This means we will need multiple passes
over classes to ensure everything gets loaded and analyzed.

Signed-off-by: Dan Heidinga <heidinga@redhat.com>